HL/#992-CAV-HH-coordination: test class added
